requirements.txt
appdirs==1.4.4
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
exceptiongroup==1.1.3
h11==0.14.0
html2image==2.0.4.3
idna==3.4
imgkit==1.2.3
importlib-metadata==6.8.0
outcome==1.2.0
Pillow==10.0.0
pycparser==2.21
pyee==8.2.2
pyppeteer==1.0.2
PySocks==1.7.1
pyTelegramBotAPI==4.13.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.4
urllib3==1.26.16
websocket-client==1.6.3
websockets==10.4
wkhtmltopdf==0.2
wsproto==1.2.0
zipp==3.16.2
